+++ src/main/java/com/example/votenow/dao/PersonDao.java
@@ -4,6 +4,6 @@ import com.example.votenow.model.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface PersonDao extends JpaRepository<Person, Long> {
+public interface PersonDao extends JpaRepository<Person, Integer> {
 }
